// src/lib/rank-assets.generated.ts
export const RANK_WINGS: Record<string, {src: string; width: number; height: number; viewBox: string; avatarLeft: number; avatarTop: number; avatarSize: number}> = {
  "BRONZE": {
    "src": "/rank-icons/wings/bronze.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "177.5 619 803 803",
    "avatarLeft": 25.965130759651306,
    "avatarTop": 10.834371108343712,
    "avatarSize": 47.820672478206724
  },
  "CHALLENGER": {
    "src": "/rank-icons/wings/challenger.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "-140.5 -28 1433 1433",
    "avatarLeft": 36.74110258199581,
    "avatarTop": 51.22121423586881,
    "avatarSize": 26.796929518492675
  },
  "DIAMOND": {
    "src": "/rank-icons/wings/diamond.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "75.5 422.5 1005 1005",
    "avatarLeft": 30.895522388059703,
    "avatarTop": 28.2089552238806,
    "avatarSize": 38.2089552238806
  },
  "EMERALD": {
    "src": "/rank-icons/wings/emerald.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "98 468.5 964 964",
    "avatarLeft": 29.87551867219917,
    "avatarTop": 24.636929460580916,
    "avatarSize": 39.83402489626556
  },
  "GOLD": {
    "src": "/rank-icons/wings/gold.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "156.5 538.5 841 841",
    "avatarLeft": 27.288941736028537,
    "avatarTop": 19.916765755053508,
    "avatarSize": 45.65992865636147
  },
  "GRANDMASTER": {
    "src": "/rank-icons/wings/grandmaster.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "2.5 262.5 1151 1151",
    "avatarLeft": 33.31885317115552,
    "avatarTop": 38.531711555169416,
    "avatarSize": 33.362293657688966
  },
  "IRON": {
    "src": "/rank-icons/wings/iron.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "238.5 666.5 686 686",
    "avatarLeft": 21.501457725947525,
    "avatarTop": 5.75801749271137,
    "avatarSize": 55.97667638483965
  },
  "MASTER": {
    "src": "/rank-icons/wings/master.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "90.5 429 977 977",
    "avatarLeft": 30.24564994882293,
    "avatarTop": 28.35209825997953,
    "avatarSize": 39.30399181166837
  },
  "PLATINUM": {
    "src": "/rank-icons/wings/platinum.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "106 521 936 936",
    "avatarLeft": 29.914529914529915,
    "avatarTop": 19.764957264957264,
    "avatarSize": 41.02564102564102
  },
  "SILVER": {
    "src": "/rank-icons/wings/silver.png",
    "width": 1156,
    "height": 1499,
    "viewBox": "137.5 573 887 887",
    "avatarLeft": 28.01578354002255,
    "avatarTop": 14.994363021420517,
    "avatarSize": 43.291995490417136
  }
};